Modified the time intervals (in seconds) to give a larger grace period to the warning and to the shutdown; this is so as to allow for longer times to unmount filesystems.
Wrapped the shutdown commands to logic that only performs this when the
system is not charging. (Not doing this will make the system shut down when
plugged into power with really low battery, like when the system is started
and this code is executed by /etc/rc.d/rc.local at startup.)

Enabled the shutdown executions to make this code fully functional.
IN 2020/09/29 "master"